Workspace/23-opspawn/4-v2/ops-core/requirements.txt
fastapi>=0.100.0,<0.112.0
uvicorn[standard]>=0.20.0,<0.30.0
pytest>=7.0.0,<9.0.0
python-dotenv>=1.0.0,<2.0.0 # For loading .env files
PyYAML>=6.0,<7.0 # For parsing YAML workflow templates
pytest-asyncio>=0.21.0,<0.24.0 # For testing async code with pytest
pytest-mock>=3.10.0,<4.0.0 # For mocking dependencies in tests
pytest-cov>=3.0.0,<6.0.0 # For code coverage reporting
python-multipart>=0.0.5,<0.0.10 # Optional FastAPI dependency for form parsing
httpx>=0.24.0,<0.28.0 # For making HTTP requests (sync/async) in SDK, tests, and agentkit_client
pytest-httpx>=0.29.0,<1.0.0 # For mocking httpx requests in tests
click>=8.0.0,<9.0.0 # For building the CLI application